cast to int16

+++ preprocess.py
@@ -48,11 +48,12 @@ def file_to_vec(audio_processor, model_settings, filename=None):
         filename, model_settings, sess
     )
 
-    return results[0]
+    return tf.cast(results[0], tf.int16)
 
 
 def get_data(audio_processor, model_settings, sample=False):
     from conf import default_conf
+
     pickle_file = default_conf['dataset_path']
     try:
         # load existing preprocessed data
